Fix tile centering

+++ src/haven/MapView.java
@@ -1734,8 +1734,8 @@ public class MapView extends PView implements DTarget, Console.Directory, PFList
 
             if (inf == null) {
                 if (Config.tilecenter && clickb == 1) {
-                    mc.x = mc.x / 11 * 11 + Integer.signum((int)mc.x) * 5;
-                    mc.y = mc.y / 11 * 11 + Integer.signum((int)mc.y) * 5;
+                    mc.x = ((int)mc.x / 11) * 11 + Integer.signum((int)mc.x) * 5;
+                    mc.y = ((int)mc.y / 11) * 11 + Integer.signum((int)mc.y) * 5;
                 }
 
                 if (Config.pf && clickb == 1 && curs != null && !curs.name.equals("gfx/hud/curs/study")) {
@@ -2011,8 +2011,8 @@ public class MapView extends PView implements DTarget, Console.Directory, PFList
             public void hit(Coord pc, Coord2d mc, ClickInfo inf) {
                 if (inf == null) {
                     if (Config.tilecenter) {
-                        mc.x = mc.x / 11 * 11 + Integer.signum((int)mc.x) * 5;
-                        mc.y = mc.y / 11 * 11 + Integer.signum((int)mc.y) * 5;
+                        mc.x = ((int)mc.x / 11) * 11 + Integer.signum((int)mc.x) * 5;
+                        mc.y = ((int)mc.y / 11) * 11 + Integer.signum((int)mc.y) * 5;
                     }
                     wdgmsg("itemact", pc, mc.floor(posres), ui.modflags());
                 } else {
